update import csv siswa (belum di test)

+++ app/Http/Controllers/SiswaController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Exports\SiswaExport;
 use App\Http\Requests\CreateSiswaRequest;
 use App\Http\Requests\Siswa\StoreRequest;
 use App\Http\Requests\Siswa\UpdateRequest;
@@ -9,16 +10,30 @@ use App\Http\Requests\UpdateSiswaRequest;
 use App\Models\Siswa;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\DB;
+use Maatwebsite\Excel\Facades\Excel;
 
 class SiswaController extends Controller
 {
+    /**
+     * index
+     *
+     * @return void
+     */
     public function index()
     {
-        $dataSiswa = Siswa::latest()->get();
+        $dataSiswa = Siswa::query()
+            ->latest()
+            ->get();
         return view('siswa.index', compact('dataSiswa'));
     }
 
+    /**
+     * create
+     *
+     * @param  mixed $request
+     * @return void
+     */
     public function create(CreateSiswaRequest $request)
     {
         try {
@@ -34,12 +49,14 @@ class SiswaController extends Controller
                 'phone_number' => $request->phone_number,
                 'nik' => $request->nik,
                 'gender' => $request->gender,
-                'image' => $nameImage
+                'image' => $nameImage,
             ];
 
             Siswa::create($data);
 
-            $dataSiswa = Siswa::latest()->get();
+            $dataSiswa = Siswa::query()
+                ->latest()
+                ->get();
 
             return response()->json(['success' => 'Berhasil', 'dataSiswa' => $dataSiswa], 200);
         } catch (\Throwable $e) {
@@ -48,16 +65,31 @@ class SiswaController extends Controller
         }
     }
 
+    /**
+     * edit
+     *
+     * @param  mixed $siswa
+     * @return void
+     */
     public function edit(Siswa $siswa)
     {
         return response()->json(['success' => 'Berhasil', 'dataSiswa' => $siswa], 200);
     }
 
+    /**
+     * update
+     *
+     * @param  mixed $request
+     * @param  mixed $siswa
+     * @return void
+     */
     public function update(UpdateSiswaRequest $request, Siswa $siswa)
     {
         try {
             if ($request->hasFile('image')) {
-                Storage::disk('public')->delete($siswa->image);
+                if ($siswa->image != null) {
+                    Storage::disk('public')->delete($siswa->image);
+                }
                 $nameImage = $request->image->store('foto-siswa', 'public');
             } else {
                 $nameImage = $siswa->image;
@@ -70,12 +102,14 @@ class SiswaController extends Controller
                 'phone_number' => $request->phone_number,
                 'nik' => $request->nik,
                 'gender' => $request->gender,
-                'image' => $nameImage
+                'image' => $nameImage,
             ];
 
             $siswa->update($data);
 
-            $dataSiswa = Siswa::latest()->get();
+            $dataSiswa = Siswa::query()
+                ->latest()
+                ->get();
 
             return response()->json(['success' => 'Berhasil', 'dataSiswa' => $dataSiswa], 200);
         } catch (\Throwable $e) {
@@ -84,15 +118,62 @@ class SiswaController extends Controller
         }
     }
 
+    /**
+     * destroy
+     *
+     * @param  mixed $siswa
+     * @return void
+     */
     public function destroy(Siswa $siswa)
     {
         $siswa->delete();
-        $dataSiswa = Siswa::latest()->get();
+        $dataSiswa = Siswa::query()
+            ->latest()
+            ->get();
 
         if ($dataSiswa->isEmpty()) {
             return response()->json(['success' => 'Berhasil', 'dataSiswa' => null], 200);
         }
 
         return response()->json(['success' => 'Berhasil', 'dataSiswa' => $dataSiswa], 200);
+    }
+
+    /**
+     * importCsv
+     * Ini adalah function untuk Import data siswa menggunakan file excel berformat csv
+     *
+     * @param  mixed $request
+     * @return void
+     */
+    public function importCsv(Request $request)
+    {
+        DB::beginTransaction();
+        try {
+            $path = $request->import->getRealPath();
+            $data = array_map('str_getcsv', file($path));
+            array_shift($data);
+            $data = array_filter($data, function ($item) {
+                return !empty(trim(implode('', $item)));
+            });
+
+            foreach ($data as $row) {
+                Siswa::create([
+                    'name' => $row[0],
+                    'class' => $row[1],
+                    'phone_number' => $row[2],
+                    'nik' => $row[3],
+                    'gender' => $row[4],
+                ]);
+            }
+
+            $siswa = Siswa::query()
+                ->latest()
+                ->get();
+
+            return response()->json(['success' => 'Berhasil import csv', 'dataSiswa' => $siswa]);
+        } catch (\Throwable $th) {
+            DB::rollBack();
+            return response()->json(['error' => 'Gagal import csv' . $th->getMessage()]);
+        }
     }
 }
